fix: move MapComponent to fix SSR issues
